Move TrafficManager and *state to userd_trafficmgr
This lets the `TrafficManager` interface live in the same package as
the one and only implementation of that interface.
